+++ src/main/java/com/hbm/entity/logic/EntityNukeExplosionMK3.java
@@ -52,6 +52,7 @@ public class EntityNukeExplosionMK3 extends EntityExplosionChunkloading {
 
 	@Override
 	protected void readEntityFromNBT(NBTTagCompound nbt) {
+		markChunkLoaderRestoredFromNBT();
 		age = nbt.getInteger("age");
 		destructionRange = nbt.getInteger("destructionRange");
 		speed = nbt.getInteger("speed");